+++ src/lib/events/handlers/json.ts
@@ -10,12 +10,13 @@ import fsp from 'fs/promises'
 // Types ============================================================
 
 type Reviver = Parameters<typeof JSON.parse>["1"]
+type Replacer = (this: any, key: string, value: any) => any
 
 // Handlers =========================================================
 
 /**
- * Reads out information saved on the event object under the name specified by the`key`
- * argument and updates it with the parsed JSON object.
+ * Reads out information saved on the event object under the name specified by the `key`
+ * parameter and updates it with the parsed JSON object.
  *
  * Alias for `JSON.parse`
  */
@@ -37,6 +38,41 @@ export function parse<Event extends E.SoybeanEvent = E.SoybeanEvent>(key: string
             if (['undefined', 'string'].includes(typeof saveOrReviver)) {
                 parsedValue = JSON.parse(initialValue, reviver)
                 e.set((saveOrReviver || key) as string, parsedValue)
+            }
+
+            end(null)
+            
+        }
+        catch (error) {
+            end(error as Error)
+        }
+    })
+}
+
+/**
+ * Reads out an object or an array saved on the event object under the name specified by the `key`
+ * parameter and updates it with the stringified value.
+ *
+ * Alias for `JSON.parse`
+ */
+export function stringify<Event extends E.SoybeanEvent = E.SoybeanEvent>(key: string, replacer?: Replacer): E.EventHandler<Event>
+export function stringify<Event extends E.SoybeanEvent = E.SoybeanEvent>(key: string, saveTo?: string, replacer?: Replacer): E.EventHandler<Event>
+export function stringify<Event extends E.SoybeanEvent = E.SoybeanEvent>(key: string, saveOrReplacer?: string | Replacer, replacer?: Replacer): E.EventHandler<Event> {
+    return (e) => new Promise<null | Error>(async end => {
+        try {
+            
+            helpers.getLoggerType(e.source)(`json.stringify "${key}"`)
+
+            let initialValue: string = e.get(key)
+            let parsedValue: any
+            
+            if (typeof saveOrReplacer === 'function') {
+                parsedValue = JSON.stringify(initialValue, saveOrReplacer)
+                e.set(key, parsedValue)
+            }
+            if (['undefined', 'string'].includes(typeof saveOrReplacer)) {
+                parsedValue = JSON.stringify(initialValue, replacer)
+                e.set((saveOrReplacer || key) as string, parsedValue)
             }
 
             end(null)