+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 ### Added
 - Added `event.stopPropagation` function allowing to half the event inside a group handler.
 - Added `json.parse` handler.
+- Added `json.stringify` handler.
 
 ### Fixed
 - Fixed a typo in the error message shown from the help command when the user tries to show details about a non-existent command.